remove PyTorchRayEstimatorWrapper and directly use PyTorchRayEstimator
