+++ python/orca/src/bigdl/orca/learn/pytorch/pytorch_ray_estimator.py
@@ -28,6 +28,9 @@ from bigdl.orca.learn.utils import maybe_dataframe_to_xshards, dataframe_to_xsha
     convert_predict_xshards_to_dataframe, update_predict_xshards, \
     process_xshards_of_pandas_dataframe
 from bigdl.orca.ray import RayContext
+from bigdl.orca.learn.ray_estimator import Estimator as OrcaRayEstimator
+from bigdl.dllib.utils.file_utils import enable_multi_fs_load, enable_multi_fs_save
+
 
 import ray
 from ray.exceptions import RayActorError
@@ -87,7 +90,7 @@ def partition_refs_to_creator(partition_refs):
     return data_creator
 
 
-class PyTorchRayEstimator:
+class PyTorchRayEstimator(OrcaRayEstimator):
     def __init__(
             self,
             *,
@@ -103,6 +106,9 @@ class PyTorchRayEstimator:
             use_tqdm=False,
             backend="torch_distributed",
             workers_per_node=1):
+        if config is not None and "batch_size" in config:
+            raise Exception("Please do not specify batch_size in config. Input batch_size in the"
+                            " fit/evaluate/predict function of the estimator instead.")
 
         # todo remove ray_ctx to run on workers
         ray_ctx = RayContext.get()
@@ -182,18 +188,45 @@ class PyTorchRayEstimator:
                             "values of backend, but got {}".format(backend))
         self.num_workers = len(self.remote_workers)
 
-    def train(self,
-              data,
-              epochs=1,
-              batch_size=32,
-              profile=False,
-              reduce_results=True,
-              info=None,
-              feature_cols=None,
-              label_cols=None):
+    def fit(self,
+            data,
+            epochs=1,
+            batch_size=32,
+            profile=False,
+            reduce_results=True,
+            info=None,
+            feature_cols=None,
+            label_cols=None):
         """
-        See the documentation in
-        'bigdl.orca.learn.pytorch.estimator.PyTorchRayEstimatorWrapper.fit'.
+        Trains a PyTorch model given training data for several epochs.
+        Calls `TrainingOperator.train_epoch()` on N parallel workers simultaneously
+        underneath the hood.
+
+        :param data: An instance of SparkXShards, a Spark DataFrame or a function that
+               takes config and batch_size as argument and returns a PyTorch DataLoader for
+               training.
+        :param epochs: The number of epochs to train the model. Default is 1.
+        :param batch_size: The number of samples per batch for each worker. Default is 32.
+               The total batch size would be workers_per_node*num_nodes.
+               If your training data is a function, you can set batch_size to be the input
+               batch_size of the function for the PyTorch DataLoader.
+        :param profile: Boolean. Whether to return time stats for the training procedure.
+               Default is False.
+        :param reduce_results: Boolean. Whether to average all metrics across all workers into
+               one dict. If a metric is a non-numerical value (or nested dictionaries), one value
+               will be randomly selected among the workers. If False, returns a list of dicts for
+               all workers.
+               Default is True.
+        :param info: An optional dictionary that can be passed to the TrainingOperator for
+               train_epoch and train_batch.
+        :param feature_cols: feature column names if data is Spark DataFrame.
+        :param label_cols: label column names if data is Spark DataFrame.
+
+        :return: A list of dictionary of metrics for every training epoch. If reduce_results is
+                False, this will return a nested list of metric dictionaries whose length will be
+                equal to the total number of workers.
+                You can also provide custom metrics by passing in a custom training_operator_cls
+                when creating the Estimator.
         """
         from bigdl.orca.data import SparkXShards
 
@@ -237,35 +270,46 @@ class PyTorchRayEstimator:
         else:
             return epoch_stats
 
-    def _process_stats(self, worker_stats):
-        stats = {
-            "num_samples": sum(
-                stats.pop("num_samples", np.nan) for stats in worker_stats)
-        }
+    def predict(self,
+                data,
+                batch_size=32,
+                feature_cols=None,
+                profile=False):
+        """
+        Using this PyTorch model to make predictions on the data.
 
-        for stat_key in worker_stats[0]:
-            if isinstance(worker_stats[0], numbers.Number):
-                stats[stat_key] = np.nanmean(
-                    [s.get(stat_key, np.nan) for s in worker_stats])
-            else:
-                stats[stat_key] = worker_stats[0][stat_key]
-        return stats
-
-    def _train_epochs(self, data_creator, epochs=1, batch_size=32, profile=False, info=None):
-        params = dict(data_creator=data_creator, epochs=epochs,
-                      batch_size=batch_size, profile=profile, info=info)
-        remote_worker_stats = []
-        for i, w in enumerate(self.remote_workers):
-            stats = w.train_epochs.remote(**params)
-            remote_worker_stats.append(stats)
-
-        success = check_for_failure(remote_worker_stats)
-        if success:
-            return success, ray.get(remote_worker_stats)
+        :param data: An instance of SparkXShards or a Spark DataFrame
+        :param batch_size: The number of samples per batch for each worker. Default is 32.
+        :param profile: Boolean. Whether to return time stats for the training procedure.
+               Default is False.
+        :param feature_cols: feature column names if data is a Spark DataFrame.
+        :return: A SparkXShards that contains the predictions with key "prediction" in each shard
+        """
+        from bigdl.orca.data import SparkXShards
+        param = dict(
+            batch_size=batch_size,
+            profile=profile
+        )
+        from pyspark.sql import DataFrame
+        if isinstance(data, DataFrame):
+            xshards, _ = dataframe_to_xshards(data,
+                                              validation_data=None,
+                                              feature_cols=feature_cols,
+                                              label_cols=None,
+                                              mode="predict")
+            pred_shards = self._predict_spark_xshards(xshards, param)
+            result = convert_predict_xshards_to_dataframe(data, pred_shards)
+        elif isinstance(data, SparkXShards):
+            if data._get_class_name() == 'pandas.core.frame.DataFrame':
+                data = process_xshards_of_pandas_dataframe(data, feature_cols)
+            pred_shards = self._predict_spark_xshards(data, param)
+            result = update_predict_xshards(data, pred_shards)
         else:
-            return success, None
+            raise ValueError("Only xshards or Spark DataFrame is supported for predict")
 
-    def validate(self,
+        return result
+
+    def evaluate(self,
                  data,
                  batch_size=32,
                  num_steps=None,
@@ -274,8 +318,31 @@ class PyTorchRayEstimator:
                  feature_cols=None,
                  label_cols=None):
         """
-        See the documentation in
-        'bigdl.orca.learn.pytorch.estimator.PyTorchRayEstimatorWrapper.evaluate'.
+        Evaluates a PyTorch model given validation data.
+        Note that only accuracy for classification with zero-based label is supported by
+        default. You can override validate_batch in TrainingOperator for other metrics.
+        Calls `TrainingOperator.validate()` on N parallel workers simultaneously
+        underneath the hood.
+
+        :param data: An instance of SparkXShards, a Spark DataFrame or a function that
+               takes config and batch_size as argument and returns a PyTorch DataLoader for
+               validation.
+        :param batch_size: The number of samples per batch for each worker. Default is 32.
+               The total batch size would be workers_per_node*num_nodes.
+               If your validation data is a function, you can set batch_size to be the input
+               batch_size of the function for the PyTorch DataLoader.
+        :param num_steps: The number of batches to compute the validation results on. This
+               corresponds to the number of times `TrainingOperator.validate_batch` is called.
+        :param profile: Boolean. Whether to return time stats for the training procedure.
+               Default is False.
+        :param info: An optional dictionary that can be passed to the TrainingOperator
+               for validate.
+        :param feature_cols: feature column names if train data is Spark DataFrame.
+        :param label_cols: label column names if train data is Spark DataFrame.
+
+        :return: A dictionary of metrics for the given data, including validation accuracy and loss.
+                You can also provide custom metrics by passing in a custom training_operator_cls
+                when creating the Estimator.
         """
         from bigdl.orca.data import SparkXShards
         data, _ = maybe_dataframe_to_xshards(data,
@@ -309,102 +376,46 @@ class PyTorchRayEstimator:
             worker_stats = ray.get([w.validate.remote(**params) for w in self.remote_workers])
         return self._process_stats(worker_stats)
 
-    def _predict_spark_xshards(self, xshards, param):
-        ray_xshards = RayXShards.from_spark_xshards(xshards)
-
-        def transform_func(worker, shards_ref):
-            data_creator = lambda config, batch_size: shards_ref
-            return worker.predict.remote(
-                data_creator, **param)
-
-        pred_shards = ray_xshards.transform_shards_with_actors(self.remote_workers,
-                                                               transform_func)
-        spark_xshards = pred_shards.to_spark_xshards()
-        return spark_xshards
-
-    def predict(self,
-                data,
-                batch_size=32,
-                feature_cols=None,
-                profile=False):
-        from bigdl.orca.data import SparkXShards
-        param = dict(
-            batch_size=batch_size,
-            profile=profile
-        )
-        from pyspark.sql import DataFrame
-        if isinstance(data, DataFrame):
-            xshards, _ = dataframe_to_xshards(data,
-                                              validation_data=None,
-                                              feature_cols=feature_cols,
-                                              label_cols=None,
-                                              mode="predict")
-            pred_shards = self._predict_spark_xshards(xshards, param)
-            result = convert_predict_xshards_to_dataframe(data, pred_shards)
-        elif isinstance(data, SparkXShards):
-            if data._get_class_name() == 'pandas.core.frame.DataFrame':
-                data = process_xshards_of_pandas_dataframe(data, feature_cols)
-            pred_shards = self._predict_spark_xshards(data, param)
-            result = update_predict_xshards(data, pred_shards)
-        else:
-            raise ValueError("Only xshards or Spark DataFrame is supported for predict")
-
-        return result
-
     def get_model(self):
-        """Returns the learned model(s)."""
+        """
+        Returns the learned PyTorch model.
+
+        :return: The learned PyTorch model.
+        """
         state = self.get_state_dict()
         model = self.model_creator(self.config)
         model_state = state["models"][0]
         model.load_state_dict(model_state)
         return model.module if hasattr(model, "module") else model
 
+    @enable_multi_fs_save
     def save(self, model_path):
-        """Saves the Estimator state to the provided model_path.
+        """
+        Saves the Estimator state (including model and optimizer) to the provided model_path.
 
         :param model_path: (str) Path to save the model.
+        :return:
         """
         state_dict = self.get_state_dict()
         torch.save(state_dict, model_path)
         return model_path
 
-    def get_state_dict(self):
-        stream_ids = [
-            worker.state_stream.remote()
-            for worker in self.remote_workers
-        ]
-        # get the first task id that finished executing.
-        [stream_id], stream_ids = ray.wait(stream_ids, num_returns=1, timeout=None)
-        byte_obj = ray.get(stream_id)
-        _buffer = io.BytesIO(byte_obj)
-        state_dict = torch.load(
-            _buffer,
-            map_location="cpu")
-        return state_dict
-
+    @enable_multi_fs_load
     def load(self, model_path):
-        """Loads the Estimator and all workers from the provided model_path.
+        """
+        Loads the Estimator state (including model and optimizer) from the provided model_path.
 
         :param model_path: (str) Path to the existing model.
         """
         state_dict = torch.load(model_path)
         self.load_state_dict(state_dict)
 
-    def load_state_dict(self, state_dict, blocking=True):
-        _buffer = io.BytesIO()
-        torch.save(state_dict, _buffer)
-        state_stream = _buffer.getvalue()
-        state_id = ray.put(state_stream)
-
-        remote_calls = [
-            worker.load_state_stream.remote(state_id)
-            for worker in self.remote_workers
-        ]
-        if blocking:
-            ray.get(remote_calls)
-
     def shutdown(self, force=False):
-        """Shuts down workers and releases resources."""
+        """
+        Shuts down workers and releases resources.
+
+        :return:
+        """
         if not force:
             cleanup = [
                 worker.shutdown.remote() for worker in self.remote_workers
@@ -428,3 +439,71 @@ class PyTorchRayEstimator:
                 ray.kill(worker)
 
         self.remote_workers = []
+
+    def _process_stats(self, worker_stats):
+        stats = {
+            "num_samples": sum(
+                stats.pop("num_samples", np.nan) for stats in worker_stats)
+        }
+
+        for stat_key in worker_stats[0]:
+            if isinstance(worker_stats[0], numbers.Number):
+                stats[stat_key] = np.nanmean(
+                    [s.get(stat_key, np.nan) for s in worker_stats])
+            else:
+                stats[stat_key] = worker_stats[0][stat_key]
+        return stats
+
+    def _train_epochs(self, data_creator, epochs=1, batch_size=32, profile=False, info=None):
+        params = dict(data_creator=data_creator, epochs=epochs,
+                      batch_size=batch_size, profile=profile, info=info)
+        remote_worker_stats = []
+        for i, w in enumerate(self.remote_workers):
+            stats = w.train_epochs.remote(**params)
+            remote_worker_stats.append(stats)
+
+        success = check_for_failure(remote_worker_stats)
+        if success:
+            return success, ray.get(remote_worker_stats)
+        else:
+            return success, None
+
+    def _predict_spark_xshards(self, xshards, param):
+        ray_xshards = RayXShards.from_spark_xshards(xshards)
+
+        def transform_func(worker, shards_ref):
+            data_creator = lambda config, batch_size: shards_ref
+            return worker.predict.remote(
+                data_creator, **param)
+
+        pred_shards = ray_xshards.transform_shards_with_actors(self.remote_workers,
+                                                               transform_func)
+        spark_xshards = pred_shards.to_spark_xshards()
+        return spark_xshards
+
+    def get_state_dict(self):
+        stream_ids = [
+            worker.state_stream.remote()
+            for worker in self.remote_workers
+        ]
+        # get the first task id that finished executing.
+        [stream_id], stream_ids = ray.wait(stream_ids, num_returns=1, timeout=None)
+        byte_obj = ray.get(stream_id)
+        _buffer = io.BytesIO(byte_obj)
+        state_dict = torch.load(
+            _buffer,
+            map_location="cpu")
+        return state_dict
+
+    def load_state_dict(self, state_dict, blocking=True):
+        _buffer = io.BytesIO()
+        torch.save(state_dict, _buffer)
+        state_stream = _buffer.getvalue()
+        state_id = ray.put(state_stream)
+
+        remote_calls = [
+            worker.load_state_stream.remote(state_id)
+            for worker in self.remote_workers
+        ]
+        if blocking:
+            ray.get(remote_calls)
